// src/domain/enums/error-message.enum.ts
export enum ErrorMessage {
  ACCOUNT_TYPE_CANNOT_BE_SAME = 'Account type cannot be the same',
  ACCOUNT_DOES_NOT_BELONG_TO_CUSTOMER = 'Account does not belong to customer',
  BANK_ACCOUNT_NOT_FOUND = 'Bank account not found',
  CUSTOMER_ALREADY_EXISTS = 'Customer already exists',
  CUSTOMER_NOT_FOUND = 'Customer not found',
  FUNCTIONARY_ALREADY_EXISTS = 'Functionary already exists',
  INVALID_CPF = 'Invalid cpf',
  INVALID_CNPJ = 'Invalid cnpj',
  INVALID_AMOUNT_OR_INSIFICIENT_BALANCE = 'Invalid amount or insufficient balance',
  INVALID_DATA_FOR_ACCOUNT = 'Data does not match account type',
  INVALID_DATA_TYPE = 'Invalid data type',
  DATA_MUST_BE_PROVIDED = 'Data must be provided',
  DATE_CANNOT_BE_FUTURE = 'Provided date cannot be in the future',
  INVALID_NAME = 'Full name must not have numbers or special characters and first name must have at least two characters',
  INVALID_SALARY = "Provided salary's value can not be lower than base salary",
  INVALID_CHECK_LIMIT = "Provided check limit's value can not be lower than base check limit",
  INVALID_CELLPHONE_NUMBER = "Cellphone number must be a national number, must be provided with its DDD code and the national cellphone prefix '9'",
  INVALID_AMOUNT = 'Invalid amount',
  INVALID_DATE = 'Invalid date',
  REQUEST_ALREADY_EXISTS = 'Request already exists',
  INSUFFICIENT_FUNDS = 'Insufficient avarage income to create a checking account',
}